+++ app/models/card/identifier.rb
@@ -26,19 +26,19 @@ class Card::Identifier
 
   private
 
-  # a number which uniquely identifies both which bank the card belongs to,
-  # and whether it is a business card or a personal one. this forms part of
+  # A 1-2 digit number which uniquely identifies both which bank the card belongs
+  # to, and whether it is a business card or a personal one. this forms part of
   # the unique identifier for each card, which allows the admin to determine
   # these things about the card at a glance.
   #
-  # The bank code is determined by the bank's id, which is always an odd
-  # number. If this is a personal card, bank_number is equal to bank.id. If
-  # this is a business card, bank_number is equal to bank.id + 1.
+  # The bank code is determined by the bank's id, which is always an odd number.
+  # If this is a personal card, bank_number is equal to bank.id. If this is a
+  # business card, bank_number is equal to bank.id + 1.
   #
-  # (This numbering system is a legacy thing from before the app existed, when
-  # we still doing everything through Fieldbook, Infusionsoft etc.)
+  # (This numbering system is a legacy thing from before this app existed, when we
+  # still doing everything through Fieldbook, Infusionsoft etc.)
   def bank_code
-    "%.2d" % (@card.personal? ? @bank.id : @bank.id + 1)
+    '%.2d' % (@card.bp == "personal" ? @bank.personal_code : @bank.business_code)
   end
 
   def card_code