+++ spec/factories/alliances.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :alliance do
+    sequence(:name) { |n| %w[OneWorld StarAlliance SkyTeam][n % 3] }
+  end
+end